Lot of changes, be sure to back up before you sync
Cleaned up and restructured the code. Animations are now easier to
implement but death animation no longer works, will have to work on
that. Beginning to work on encounters and using random enemy generation
rather than hard coding it. Throwing an error right now so you won't be
able to get out of battle view. Also on occasion the game will fade out
and freeze in the middle of the battle view transition, not sure why...